Resolved route delete issue

+++ cloud/services/compute/networks/reconcile.go
@@ -18,16 +18,20 @@ package networks
 
 import (
 	"context"
-	"k8s.io/utils/pointer"
+	"strings"
 
+	"github.com/GoogleCloudPlatform/k8s-cloud-provider/pkg/cloud/filter"
 	"github.com/GoogleCloudPlatform/k8s-cloud-provider/pkg/cloud/meta"
 	"google.golang.org/api/compute/v1"
-
+	"k8s.io/utils/pointer"
 	"k8s.io/utils/ptr"
-
 	infrav1 "sigs.k8s.io/cluster-api-provider-gcp/api/v1beta1"
 	"sigs.k8s.io/cluster-api-provider-gcp/cloud/gcperrors"
 	"sigs.k8s.io/controller-runtime/pkg/log"
+)
+
+const (
+	k8sNodeRouteTag = "k8s-node-route"
 )
 
 // Reconcile reconcile cluster network components.
@@ -103,8 +107,33 @@ func (s *Service) Delete(ctx context.Context) error {
 		// Custom mode detected
 		for _, subnet := range s.scope.SubnetSpecs() {
 			subnetKey := meta.RegionalKey(subnet.Name, subnet.Region)
-			if err = s.subnets.Delete(ctx, subnetKey); err != nil {
-				log.Error(err, "Error deleting a subnetwork", "name", subnet.Name, "region", subnet.Region)
+			subnetwork, err := s.subnets.Get(ctx, subnetKey)
+			if err != nil && !gcperrors.IsNotFound(err) {
+				return err
+			}
+			if subnetwork != nil {
+				if err = s.subnets.Delete(ctx, subnetKey); err != nil {
+					log.Error(err, "Error deleting a subnetwork", "name", subnet.Name, "region", subnet.Region)
+					return err
+				}
+			}
+		}
+	}
+
+	// Delete routes associated with network
+	fl := filter.Regexp("description", k8sNodeRouteTag)
+	routeList, err := s.routes.List(ctx, fl)
+	if err != nil {
+		log.Error(err, "failed to list routes for the cluster")
+		return err
+	}
+
+	for _, route := range routeList {
+		if strings.HasSuffix(route.Network, s.scope.NetworkName()) {
+			log.V(2).Info("Deleting route ", "route:", route.Name)
+			err := s.routes.Delete(ctx, meta.GlobalKey(route.Name))
+			if err != nil {
+				log.Error(err, "Error deleting a route", "name", route.Name)
 				return err
 			}
 		}
